server/requirements.txt
# Core dependencies
pipecat-ai[silero]>=0.0.50
aiohttp>=3.9.0,<4.0.0
websockets>=13.0,<15.0
python-dotenv>=1.0.0,<2.0.0

# AI Service providers - Updated to latest Google Gen AI SDK
google-genai>=0.3.0,<1.0.0

# Google APIs for Gmail integration
google-api-python-client>=2.110.0,<3.0.0
google-auth>=2.17.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-oauth2-tool>=0.0.3

# Twilio for phone integration
twilio>=8.10.0,<9.0.0

# Web framework and utilities
fastapi[standard]>=0.115.0,<0.116.0
uvicorn[standard]>=0.24.0,<1.0.0
aiofiles>=23.2.1,<24.0.0
requests>=2.31.0,<3.0.0

# Logging and monitoring
structlog>=23.2.0,<24.0.0
uvloop>=0.19.0,<1.0.0

# Development dependencies
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=23.0.0,<25.0.0
flake8>=6.0.0,<8.0.0

# Production dependencies
gunicorn>=21.2.0,<23.0.0
prometheus-client>=0.19.0,<1.0.0